backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.28
pydantic>=2.13.0
pydantic-settings>=2.9.0
python-dotenv==1.0.0
httpx==0.28.1
passlib[argon2]==1.7.4
python-jose[cryptography]==3.5.0
APScheduler==3.10.1
python-multipart==0.0.7
